erase logs, create mood component
