--- src/front/js/component/moodRectangle.js
@@ -1,40 +1,0 @@
-import React, { useEffect, useRef } from 'react';
-
-const MoodRectangle = () => {
-    const svgRef = useRef(); // Referencia al elemento SVG
-    const foreignObjectRef = useRef(); // Referencia al elemento foreignObject
-
-    useEffect(() => {
-        const svg = svgRef.current;
-        const foreignObject = foreignObjectRef.current;
-        const div = foreignObject.querySelector('div');
-
-        // Usar una función asíncrona dentro de useEffect
-        setTimeout(() => {
-            let rectHeight = div.offsetHeight + 20; // Agrega tu padding deseado aquí
-            svg.setAttribute('height', rectHeight + 40); // Ajusta según necesites espacio para la sombra
-            foreignObject.setAttribute('height', rectHeight);
-            svg.querySelector('rect').setAttribute('height', rectHeight);
-        }, 0);
-    }, []);
-    
-    return (
-        <svg ref={svgRef} width="350" height="300" xmlns="http://www.w3.org/2000/svg">
-            <defs>
-                <filter id="shadow" x="-10%" y="-10%" width="150%" height="150%">
-                    <feDropShadow dx="10" dy="10" stdDeviation="0" floodColor="black" />
-                </filter>
-            </defs>
-            <rect x="10" y="10" width="330" height="150" rx="15" ry="15"
-                fill="#EB5F46" stroke="black" strokeWidth="4"
-                filter="url(#shadow)" />
-            <foreignObject ref={foreignObjectRef} x="10" y="10" width="330" height="150">
-                <div xmlns="http://www.w3.org/1999/xhtml" style={{ fontSize: '55px', fontFamily: 'Impact', fontWeight: '900', padding: '20px', boxSizing: 'border-box', whiteSpace: 'pre-wrap', color: '#F4B6D3', lineHeight: 1}}>
-                    ¿Para qué me he levantado de la cama?
-                </div>
-            </foreignObject>
-        </svg>
-    );
-};
-
-export default MoodRectangle;
