+++ src/front/js/pages/userProfile.js
@@ -1,4 +1,4 @@
-import React, { useState, useContext, useEffect } from "react";
+import React, { useState, useContext, useEffect,} from "react";
 import { Context } from "../store/appContext";
 import { Link, useParams } from "react-router-dom";
 import "../../styles/profile.css";
@@ -18,19 +18,24 @@ import Form from 'react-bootstrap/Form';
 
 export const UserProfile = () => {
     const { store, actions } = useContext(Context);
+    const params = useParams();
 
+    const activeUser = store?.active_users.find(user => user.id === parseInt(params.uid));
+
+    console.log(activeUser);
 
     return (
+
         <Container fluid className="container-landingpage">
             <Container className="user-profile pb-4 pt-4">
-                <Row className="mb-5">
+                 <Row className="mb-5">
                     <Col xs={11} md={6} lg={10}>
-                        <h2 className="">{store.user?.name} {store.user?.surnames}</h2>
-                        <h3 className="">{store.user?.hobbie}</h3>
+                        <h2 className="">{activeUser?.name} {activeUser?.surnames}</h2>
+                        <h3 className="">{activeUser?.hobbie}</h3>
                     </Col>
                     <Col xs={1} md={6} lg={2} className="text-end">
-                        {store.user?.user_mood ? (
-                            <Image src={store.user.user_mood.category_mood.icon_url} className="img-mood" />
+                        {activeUser?.user_mood ? (
+                            <Image src={activeUser?.user_mood.category_mood.icon_url} className="img-mood" />
                         ) : (<div>no hay mood</div>)}
                     </Col>
                 </Row>
@@ -38,21 +43,22 @@ export const UserProfile = () => {
                     <Col xs={12} md={6} lg={4}>
                         <Card className="user-info-profile">
                             <ListGroup variant="flush">
-                                <ListGroup.Item><small>Email</small><p>{store.user?.email}</p></ListGroup.Item>
+                                <ListGroup.Item><small>Email</small><p>{activeUser?.email}</p></ListGroup.Item>
                                 <ListGroup.Item><small>Contraseña</small><p>Cambiar contraseña <small>&#8599;</small></p></ListGroup.Item>
-                                <ListGroup.Item><small>Unido/a</small><p>{store.user?.created_at}</p></ListGroup.Item>
+                                <ListGroup.Item><small>Unido/a</small><p>{activeUser?.created_at}</p></ListGroup.Item>
                             </ListGroup>
                         </Card>
                     </Col>
                     <Col xs={12} md={6} lg={8}>
-                        <h4 className="border-bottom border-dark mb-3">Mis psicólogos de referencia</h4>
-                        {store.user?.psychologists.length === 0 ? (
+                        <Row>
+                        <h4 className="border-bottom border-dark mb-3 p-0">Mis psicólogos de referencia</h4>
+                        {activeUser?.psychologists.length === 0 ? (
                             <div>No tienes psycógolos de referencia. Cuando realices alguna sesión con ellos, se guardarán aquí.</div>
                         ) : (
-                            store.user?.psychologists.map((psychologist, index) => {
+                            activeUser?.psychologists.map((psychologist, index) => {
                                 return (
                                     <Col xs={12} md={6} lg={6} className="text-center" key={index}>
-                                        <div className="ps-card p-2">
+                                        <div className="ps-card p-3">
                                             <Image src={psychologist.profile_url} className="img-mood" roundedCircle />
                                             <h5>{psychologist.name} {psychologist.surnames}</h5>
                                             <p><small> [nº. {psychologist.collegiate_number} ]</small></p>
@@ -62,8 +68,16 @@ export const UserProfile = () => {
                                 )
                             })
                         )}
+                        </Row>
                     </Col>
                 </Row>
+
+                </Container>
+        </Container >
+    );
+};
+
+
 
                 {/* <Row className="mt-5">
                     <Col xs={12} md={12} lg={12}>
@@ -72,10 +86,10 @@ export const UserProfile = () => {
                     </Col>
                 </Row>
                 <Row>
-                    {store.user?.psychologists.length === 0 ? (
+                    {activeUser?.psychologists.length === 0 ? (
                         <div>No tienes psycógolos de referencia. Cuando realices alguna sesión con ellos, se guardarán aquí.</div>
                     ) : (
-                        store.user?.psychologists.map((psychologist, index) => {
+                        activeUser?.psychologists.map((psychologist, index) => {
                             return (
                                 <Col xs={12} md={6} lg={6} className="text-center" key={index}>
                                     <div className="ps-card p-3">
@@ -110,11 +124,6 @@ export const UserProfile = () => {
                         </ul>
                     </Col>
                 </Row> */}
-            </Container>
-        </Container >
-    );
-};
-
