+++ src/front/js/store/appContext.js
@@ -31,6 +31,7 @@ const injectContext = PassedComponent => {
 			state.actions.getMessage(); // <---- calling this function from the flux.js actions
 			state.actions.validToken();
 			state.actions.getCurrentUser();
+			state.actions.getAllActiveUsers();
 		}, []);
 
 		// The initial value for the context is not null anymore, but the current state of this component,
