+++ src/front/styles/profile.css
@@ -213,4 +213,6 @@ div .tab-content{
  .ps-card {
     border: 1px solid var(--base-color-text);
     background-color: white;
+
  }
+
